requirements.txt
# Core FastAPI and web framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# AI and ML libraries - using more flexible versions
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0

# Vector database and embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
langchain-huggingface

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# HTTP client
aiohttp>=3.8.0

# Text processing
tiktoken>=0.5.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.5.0

# Production dependencies
psutil>=5.9.0
python-json-logger>=2.0.0

# Security and middleware
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6